Isolate login and signup forms to be reused at join
